+++ src/Command/Migrator/CreateMigrationCommandFactory.php
@@ -9,7 +9,7 @@ use Psr\Container\ContainerExceptionInterface;
 use Psr\Container\ContainerInterface;
 use Psr\Container\NotFoundExceptionInterface;
 
-class CreateMigrationCommandFactory
+final class CreateMigrationCommandFactory
 {
     /**
      * @throws ContainerExceptionInterface