+++ src/Command/Migrator/RollbackCommandFactory.php
@@ -9,7 +9,7 @@ use Psr\Container\ContainerInterface;
 use Psr\Container\NotFoundExceptionInterface;
 use Sirix\Cycle\Service\MigratorService;
 
-class RollbackCommandFactory
+final class RollbackCommandFactory
 {
     /**
      * @throws ContainerExceptionInterface