+++ src/Command/Migrator/SeedCommandFactory.php
@@ -10,7 +10,7 @@ use Psr\Container\ContainerInterface;
 use Psr\Container\NotFoundExceptionInterface;
 use Sirix\Cycle\Service\MigratorService;
 
-class SeedCommandFactory
+final class SeedCommandFactory
 {
     /**
      * @throws ContainerExceptionInterface